# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handler.BatchDisableUsersRequest:
    properties:
      userIds:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - userIds
    type: object
...
  handler.CreateUserRequest:
    properties:
      avatar:
        type: string
      email:
        example: zhangsan@example.com
        type: string
      mobile:
        example: "13800138000"
        type: string
      nickname:
        example: 张三
        type: string
      parentId:
        type: integer
      password:
        example: "123456"
        maxLength: 64
        minLength: 6
        type: string
      roleId:
        example: 1
        type: integer
      username:
        example: zhangsan
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - roleId
    - username
    type: object
...
  handler.UpdateUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      mobile:
        type: string
      nickname:
        type: string
      parentId:
        type: integer
      password:
        type: string
      roleId:
        type: integer
      status:
        example: enabled
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - roleId
    - username
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      parentId:
        type: integer
      roleId:
        type: integer
      salesType:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  model.UserWithRole:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      parentId:
        type: integer
      roleId:
        type: integer
      roleLabel:
        type: string
      roleName:
        type: string
      salesType:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 后端服务API文档
  title: Backend API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserWithRole'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - users
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户详情
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - users
...
  /api/v1/users/batch/disable:
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量禁用信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.BatchDisableUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    format: int64
                    type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 批量禁用用户
      tags:
      - users
  /api/v1/users/search:
    get:
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserWithRole'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 搜索用户
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Bearer token, 格式: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"